+++ CHANGELOG.md
@@ -6,6 +6,9 @@ https://github.com/clojure/brew-install/blob/1.10.1/CHANGELOG.md
 
 ## Releases
 
+- Third release [2019-11-21]
+    - Based on clojure tools v.1.10.1.489
+
 - Second release [2019-11-11]
     - Based on clojure tools v.1.10.1.486
     - Added a "Why?" section in the readme file